move handleEdges to utils
