+++ sketches/particles/particles.ts
@@ -1,11 +1,11 @@
 import * as Vector from './Vector';
 import { type IParticle, create, applyForce } from './Particle';
-import { getCanvas, getCanvasContext, type IPointTuple, loop } from './particles.utils';
+import { getCanvas, getCanvasContext, handleEdges, type IPointTuple, loop } from './particles.utils';
 import * as Vec from './Vector';
 
-const INITIAL_SPEED = 250;
-const FPS = 60;
-const DELTA_TIME = 1 / FPS;
+export const INITIAL_SPEED = 250;
+export const FPS = 60;
+export const DELTA_TIME = 1 / FPS;
 
 document.body.onload = () => {
   const canvas = getCanvas();
@@ -29,33 +29,17 @@ interface ISketchData {
 
 const render = (context: CanvasRenderingContext2D, data: ISketchData) => (t: number) => {
   const { width, height } = context.canvas;
-
   const { particle } = data;
-
-  const force = handleEdges(particle, width, height);
 
   data.particle = applyForce({
     particle,
-    force,
+    force: handleEdges(particle, width, height),
     deltaTime: DELTA_TIME,
   });
 
-  console.log(data.particle.velocity);
-
   context.clearRect(0, 0, width, height);
-
   context.beginPath();
   context.arc(...Vec.toTuple(particle.position), 10, 0, 2 * Math.PI);
   context.fillStyle = 'black';
   context.fill();
 };
-
-function handleEdges(particle: IParticle, width: number, height: number) {
-  const { position, velocity, mass } = particle;
-
-  const force = Vector.create(
-    position.x < 0 || position.x > width ? (-2 * velocity.x * mass) / DELTA_TIME : 0,
-    position.y < 0 || position.y > height ? (-2 * velocity.y * mass) / DELTA_TIME : 0
-  );
-  return force;
-}
